@@ -26,7 +26,7 @@ module.exports = {
     label: "sdk",
     algolia: {
       id: "QLS2QSP47E",
-      key: "067b84458bfa80c295e1d4f12c461911",
+      key: "4d9feeb481e3cfef8f91bbc63e090042",
       index: "cosmos_network_vue"
     },
     versions: [
