@@ -11,6 +11,6 @@ const (
 
 	AttributeKeyWithdrawAddress = "withdraw_address"
 	AttributeKeyValidator       = "validator"
-
-	AttributeValueCategory = ModuleName
+	AttributeKeyDelegator       = "delegator"
+	AttributeValueCategory      = ModuleName
 )
