@@ -1380,6 +1380,111 @@ func (m *QueryTallyResultResponse) GetTally() TallyResult {
 	return TallyResult{}
 }
 
+// QueryGroupsRequest is the Query/Groups request type.
+//
+// Since: cosmos-sdk 0.47.1
+type QueryGroupsRequest struct {
+	// pagination defines an optional pagination for the request.
+	Pagination *query.PageRequest `protobuf:"bytes,2,opt,name=pagination,proto3" json:"pagination,omitempty"`
+}
+
+func (m *QueryGroupsRequest) Reset()         { *m = QueryGroupsRequest{} }
+func (m *QueryGroupsRequest) String() string { return proto.CompactTextString(m) }
+func (*QueryGroupsRequest) ProtoMessage()    {}
+func (*QueryGroupsRequest) Descriptor() ([]byte, []int) {
+	return fileDescriptor_0fcf9f1d74302290, []int{26}
+}
+func (m *QueryGroupsRequest) XXX_Unmarshal(b []byte) error {
+	return m.Unmarshal(b)
+}
+func (m *QueryGroupsRequest) XXX_Marshal(b []byte, deterministic bool) ([]byte, error) {
+	if deterministic {
+		return xxx_messageInfo_QueryGroupsRequest.Marshal(b, m, deterministic)
+	} else {
+		b = b[:cap(b)]
+		n, err := m.MarshalToSizedBuffer(b)
+		if err != nil {
+			return nil, err
+		}
+		return b[:n], nil
+	}
+}
+func (m *QueryGroupsRequest) XXX_Merge(src proto.Message) {
+	xxx_messageInfo_QueryGroupsRequest.Merge(m, src)
+}
+func (m *QueryGroupsRequest) XXX_Size() int {
+	return m.Size()
+}
+func (m *QueryGroupsRequest) XXX_DiscardUnknown() {
+	xxx_messageInfo_QueryGroupsRequest.DiscardUnknown(m)
+}
+
+var xxx_messageInfo_QueryGroupsRequest proto.InternalMessageInfo
+
+func (m *QueryGroupsRequest) GetPagination() *query.PageRequest {
+	if m != nil {
+		return m.Pagination
+	}
+	return nil
+}
+
+// QueryGroupsResponse is the Query/Groups response type.
+//
+// Since: cosmos-sdk 0.47.1
+type QueryGroupsResponse struct {
+	// `groups` is all the groups present in state.
+	Groups []*GroupInfo `protobuf:"bytes,1,rep,name=groups,proto3" json:"groups,omitempty"`
+	// pagination defines the pagination in the response.
+	Pagination *query.PageResponse `protobuf:"bytes,2,opt,name=pagination,proto3" json:"pagination,omitempty"`
+}
+
+func (m *QueryGroupsResponse) Reset()         { *m = QueryGroupsResponse{} }
+func (m *QueryGroupsResponse) String() string { return proto.CompactTextString(m) }
+func (*QueryGroupsResponse) ProtoMessage()    {}
+func (*QueryGroupsResponse) Descriptor() ([]byte, []int) {
+	return fileDescriptor_0fcf9f1d74302290, []int{27}
+}
+func (m *QueryGroupsResponse) XXX_Unmarshal(b []byte) error {
+	return m.Unmarshal(b)
+}
+func (m *QueryGroupsResponse) XXX_Marshal(b []byte, deterministic bool) ([]byte, error) {
+	if deterministic {
+		return xxx_messageInfo_QueryGroupsResponse.Marshal(b, m, deterministic)
+	} else {
+		b = b[:cap(b)]
+		n, err := m.MarshalToSizedBuffer(b)
+		if err != nil {
+			return nil, err
+		}
+		return b[:n], nil
+	}
+}
+func (m *QueryGroupsResponse) XXX_Merge(src proto.Message) {
+	xxx_messageInfo_QueryGroupsResponse.Merge(m, src)
+}
+func (m *QueryGroupsResponse) XXX_Size() int {
+	return m.Size()
+}
+func (m *QueryGroupsResponse) XXX_DiscardUnknown() {
+	xxx_messageInfo_QueryGroupsResponse.DiscardUnknown(m)
+}
+
+var xxx_messageInfo_QueryGroupsResponse proto.InternalMessageInfo
+
+func (m *QueryGroupsResponse) GetGroups() []*GroupInfo {
+	if m != nil {
+		return m.Groups
+	}
+	return nil
+}
+
+func (m *QueryGroupsResponse) GetPagination() *query.PageResponse {
+	if m != nil {
+		return m.Pagination
+	}
+	return nil
+}
+
 func init() {
 	proto.RegisterType((*QueryGroupInfoRequest)(nil), "cosmos.group.v1.QueryGroupInfoRequest")
 	proto.RegisterType((*QueryGroupInfoResponse)(nil), "cosmos.group.v1.QueryGroupInfoResponse")
@@ -1407,91 +1512,95 @@ func init() {
 	proto.RegisterType((*QueryGroupsByMemberResponse)(nil), "cosmos.group.v1.QueryGroupsByMemberResponse")
 	proto.RegisterType((*QueryTallyResultRequest)(nil), "cosmos.group.v1.QueryTallyResultRequest")
 	proto.RegisterType((*QueryTallyResultResponse)(nil), "cosmos.group.v1.QueryTallyResultResponse")
+	proto.RegisterType((*QueryGroupsRequest)(nil), "cosmos.group.v1.QueryGroupsRequest")
+	proto.RegisterType((*QueryGroupsResponse)(nil), "cosmos.group.v1.QueryGroupsResponse")
 }
 
 func init() { proto.RegisterFile("cosmos/group/v1/query.proto", fileDescriptor_0fcf9f1d74302290) }
 
 var fileDescriptor_0fcf9f1d74302290 = []byte{
-	// 1249 bytes of a gzipped FileDescriptorProto
-	0x1f, 0x8b, 0x08, 0x00, 0x00, 0x00, 0x00, 0x00, 0x02, 0xff, 0xcc, 0x98, 0xcf, 0x6f, 0xdc, 0x44,
-	0x14, 0xc7, 0x33, 0x25, 0x69, 0x92, 0x97, 0xb6, 0x11, 0x43, 0x5a, 0x12, 0x37, 0xda, 0x04, 0x03,
-	0xf9, 0x1d, 0x3b, 0xbb, 0x49, 0xd3, 0x8a, 0x9f, 0xea, 0x4a, 0x10, 0x72, 0x28, 0x4a, 0x97, 0x88,
-	0x03, 0x97, 0xc8, 0x9b, 0x75, 0x8c, 0xc5, 0xae, 0x67, 0xbb, 0x76, 0x22, 0x56, 0xd1, 0x5e, 0x90,
-	0xca, 0x01, 0x71, 0x00, 0x8a, 0x50, 0x89, 0x38, 0xf4, 0x80, 0x04, 0x7f, 0x00, 0x08, 0x89, 0x5b,
-	0x6f, 0x3d, 0x56, 0x70, 0xe1, 0x84, 0x50, 0xc2, 0x1f, 0x82, 0x3c, 0xf3, 0xbc, 0xeb, 0xdf, 0xeb,
-	0x15, 0x2b, 0xc8, 0xa9, 0x5d, 0xfb, 0xbd, 0x79, 0x9f, 0xf9, 0xbe, 0xe7, 0xf1, 0xd7, 0x81, 0xeb,
-	0xfb, 0xcc, 0xae, 0x31, 0x5b, 0x35, 0x1a, 0xec, 0xb0, 0xae, 0x1e, 0xe5, 0xd5, 0x7b, 0x87, 0x7a,
-	0xa3, 0xa9, 0xd4, 0x1b, 0xcc, 0x61, 0x74, 0x5c, 0xdc, 0x54, 0xf8, 0x4d, 0xe5, 0x28, 0x2f, 0x4d,
-	0x18, 0xcc, 0x60, 0xfc, 0x9e, 0xea, 0xfe, 0x4f, 0x84, 0x49, 0xd3, 0x06, 0x63, 0x46, 0x55, 0x57,
-	0xb5, 0xba, 0xa9, 0x6a, 0x96, 0xc5, 0x1c, 0xcd, 0x31, 0x99, 0x65, 0xe3, 0xdd, 0x48, 0x05, 0xa7,
-	0x59, 0xd7, 0xbd, 0x9b, 0x4b, 0x78, 0xb3, 0xac, 0xd9, 0xba, 0x28, 0xad, 0x1e, 0xe5, 0xcb, 0xba,
-	0xa3, 0xe5, 0xd5, 0xba, 0x66, 0x98, 0x16, 0x5f, 0x09, 0x63, 0xa7, 0x44, 0xec, 0x9e, 0xa8, 0x8f,
-	0x68, 0xfc, 0x87, 0x5c, 0x80, 0xab, 0x77, 0xdd, 0xe4, 0x2d, 0xb7, 0xc6, 0xb6, 0x75, 0xc0, 0x4a,
-	0xfa, 0xbd, 0x43, 0xdd, 0x76, 0xe8, 0x14, 0x8c, 0xf0, 0xba, 0x7b, 0x66, 0x65, 0x92, 0xcc, 0x92,
-	0x85, 0xc1, 0xd2, 0x30, 0xff, 0xbd, 0x5d, 0x91, 0xdf, 0x81, 0x6b, 0xe1, 0x1c, 0xbb, 0xce, 0x2c,
-	0x5b, 0xa7, 0x0a, 0x0c, 0x9a, 0xd6, 0x01, 0xe3, 0x09, 0x63, 0x05, 0x49, 0x09, 0xa9, 0xa0, 0x74,
-	0x32, 0x78, 0x9c, 0x7c, 0x17, 0xae, 0x77, 0x56, 0xda, 0x61, 0x55, 0x73, 0xbf, 0xe9, 0x67, 0x28,
-	0xc0, 0xb0, 0x56, 0xa9, 0x34, 0x74, 0xdb, 0xe6, 0x2b, 0x8e, 0x16, 0x27, 0x7f, 0xfb, 0x69, 0x75,
-	0x02, 0x17, 0xbd, 0x2d, 0xee, 0xbc, 0xe7, 0x34, 0x4c, 0xcb, 0x28, 0x79, 0x81, 0xf2, 0x2e, 0x4c,
-	0xc7, 0x2f, 0x89, 0x88, 0x1b, 0x01, 0xc4, 0xd9, 0x78, 0x44, 0x5f, 0x9e, 0x00, 0x6d, 0xc1, 0x64,
-	0x67, 0xd5, 0x3b, 0x7a, 0xad, 0xac, 0x37, 0xec, 0xee, 0x4a, 0xd1, 0xb7, 0x01, 0x3a, 0xcd, 0x98,
-	0xbc, 0xc0, 0x4b, 0xce, 0x79, 0x25, 0xdd, 0xce, 0x29, 0x62, 0x68, 0xb0, 0x73, 0xca, 0x8e, 0x66,
-	0xe8, 0xb8, 0x6c, 0xc9, 0x97, 0x29, 0x7f, 0x47, 0x60, 0x2a, 0xa6, 0x3e, 0x6e, 0x69, 0x13, 0x86,
-	0x6b, 0xe2, 0xd2, 0x24, 0x99, 0x7d, 0x66, 0x61, 0xac, 0x30, 0x1d, 0xbf, 0x2b, 0x91, 0x57, 0xf2,
-	0x82, 0xe9, 0x56, 0x0c, 0xdd, 0x7c, 0x57, 0x3a, 0x51, 0x34, 0x80, 0xf7, 0x20, 0x80, 0x67, 0x17,
-	0x9b, 0xb7, 0x2b, 0x35, 0xd3, 0xf2, 0xf4, 0x51, 0x60, 0x48, 0x73, 0x7f, 0x77, 0xed, 0xa1, 0x08,
-	0xeb, 0x9b, 0x68, 0xdf, 0x12, 0x90, 0xe2, 0xa8, 0x50, 0xb5, 0x02, 0x5c, 0xe4, 0xf2, 0x78, 0xa2,
-	0xa5, 0x4d, 0x2b, 0x46, 0xf6, 0x4f, 0xb1, 0xfb, 0x04, 0x66, 0x43, 0x63, 0x6a, 0xea, 0x76, 0x51,
-	0xfc, 0xfc, 0x0f, 0x07, 0xeb, 0x67, 0x02, 0x2f, 0xa4, 0x70, 0xa0, 0x54, 0x5b, 0x70, 0x45, 0x80,
-	0xd4, 0x31, 0x00, 0x25, 0xeb, 0xfe, 0xf4, 0x5c, 0x36, 0xfc, 0xeb, 0xf6, 0x4f, 0xbf, 0x93, 0x04,
-	0xfd, 0xce, 0xc5, 0xe0, 0x25, 0x89, 0x1a, 0x9c, 0xbf, 0xf3, 0x27, 0xea, 0x4d, 0x98, 0xe0, 0xd8,
-	0x3b, 0x0d, 0x56, 0x67, 0xb6, 0x56, 0xf5, 0x74, 0x9c, 0x81, 0xb1, 0x3a, 0x5e, 0xea, 0x8c, 0x22,
-	0x78, 0x97, 0xb6, 0x2b, 0xf2, 0xbb, 0xf8, 0x12, 0xe9, 0x24, 0xe2, 0x1e, 0x6f, 0xc0, 0x88, 0x17,
-	0x86, 0x07, 0xee, 0x54, 0x64, 0x77, 0xed, 0xa4, 0x76, 0xa8, 0xfc, 0x88, 0x80, 0x1c, 0x58, 0xd0,
-	0x9b, 0x48, 0x21, 0xc2, 0xbf, 0x78, 0x3d, 0xf4, 0xad, 0xc7, 0x3f, 0x10, 0x78, 0x31, 0x15, 0x11,
-	0x15, 0xb8, 0x09, 0xa3, 0xde, 0xb6, 0xbc, 0x06, 0xa7, 0x48, 0xd0, 0x89, 0xed, 0x5f, 0x57, 0x1b,
-	0x30, 0xc3, 0x41, 0xdf, 0x67, 0x8e, 0x5e, 0x6c, 0xe3, 0xba, 0xbf, 0x1a, 0x59, 0x1b, 0xec, 0x3e,
-	0x49, 0x47, 0x6e, 0x02, 0xe7, 0x48, 0x7d, 0x92, 0x78, 0x98, 0x7c, 0x07, 0x9f, 0xce, 0xd8, 0x9a,
-	0xa8, 0xcc, 0x22, 0x0c, 0xba, 0xc1, 0x38, 0x17, 0x57, 0x23, 0xa2, 0xb8, 0xd1, 0x25, 0x1e, 0x22,
-	0x7f, 0x4a, 0xd0, 0x27, 0xb8, 0xd7, 0xec, 0x62, 0xcf, 0x03, 0xda, 0xb7, 0xae, 0x7f, 0x4d, 0xd0,
-	0x5d, 0x44, 0x40, 0x70, 0x53, 0xcb, 0x42, 0x28, 0xaf, 0xd5, 0x09, 0xbb, 0x12, 0x31, 0xfd, 0x6b,
-	0xf1, 0x57, 0x04, 0xed, 0x09, 0x62, 0x05, 0x9a, 0xdb, 0xee, 0x1d, 0xc9, 0xd4, 0xbb, 0xbe, 0x69,
-	0xf5, 0xa5, 0x67, 0x0a, 0x82, 0x50, 0xff, 0xab, 0x50, 0x0f, 0xc3, 0x96, 0x00, 0x2d, 0xd1, 0x39,
-	0x38, 0x50, 0x4e, 0x88, 0xdf, 0x0b, 0xfb, 0xd0, 0xce, 0x83, 0x5d, 0x79, 0x05, 0x9e, 0xe7, 0x6c,
-	0xbb, 0x5a, 0xb5, 0xea, 0x9e, 0x6d, 0x87, 0x55, 0x27, 0xf3, 0xcb, 0x61, 0x17, 0x67, 0x33, 0x90,
-	0x8b, 0x9b, 0xba, 0x05, 0x43, 0x8e, 0x7b, 0x19, 0x0f, 0x81, 0xa8, 0x6f, 0xf5, 0x25, 0x15, 0x07,
-	0x9f, 0xfc, 0x39, 0x33, 0x50, 0x12, 0x09, 0x85, 0xfb, 0xcf, 0xc2, 0x10, 0x5f, 0x96, 0x7e, 0x4e,
-	0x60, 0xb4, 0xbd, 0x75, 0x3a, 0x17, 0x59, 0x22, 0xf6, 0xf3, 0x46, 0x9a, 0xef, 0x1a, 0x27, 0x10,
-	0x65, 0xe5, 0x93, 0xdf, 0xff, 0x7e, 0x70, 0x61, 0x81, 0xce, 0xa9, 0xe1, 0xaf, 0x31, 0xf4, 0x66,
-	0xd6, 0x01, 0x53, 0x8f, 0x3d, 0x9f, 0xd6, 0xa2, 0xdf, 0x13, 0x18, 0x0f, 0xbd, 0xb0, 0xe9, 0x4a,
-	0x4a, 0xb1, 0xc8, 0x57, 0x8f, 0xb4, 0x9a, 0x31, 0x1a, 0x01, 0x37, 0x38, 0xa0, 0x42, 0x57, 0x12,
-	0x00, 0xb9, 0xbd, 0x68, 0x22, 0x27, 0x4e, 0x6d, 0x8b, 0x3e, 0x24, 0x70, 0xc9, 0xff, 0x31, 0x41,
-	0x17, 0x53, 0xaa, 0x06, 0x3f, 0x78, 0xa4, 0xa5, 0x2c, 0xa1, 0x48, 0x97, 0xe7, 0x74, 0xcb, 0x74,
-	0x31, 0x81, 0x0e, 0xbf, 0x45, 0xfc, 0x0a, 0x9e, 0x10, 0xb8, 0x1c, 0xb0, 0xec, 0x34, 0xad, 0x60,
-	0xc8, 0xf4, 0x49, 0xcb, 0x99, 0x62, 0x91, 0x6e, 0x8d, 0xd3, 0x2d, 0xd1, 0x85, 0x78, 0x3a, 0x7b,
-	0xaf, 0xdc, 0xdc, 0xe3, 0xde, 0xd0, 0x55, 0xae, 0x66, 0x5a, 0x2d, 0xfa, 0x2b, 0x81, 0x89, 0x38,
-	0xaf, 0x4c, 0xf3, 0xdd, 0xba, 0x16, 0xf1, 0xf7, 0x52, 0xa1, 0x97, 0x14, 0x24, 0x7e, 0x95, 0x13,
-	0xdf, 0xa0, 0xeb, 0x69, 0xdd, 0x36, 0x75, 0x4e, 0x2e, 0x6e, 0xf9, 0x94, 0xfd, 0x25, 0x0a, 0x2f,
-	0x04, 0xce, 0x06, 0x1f, 0xd0, 0xb9, 0xd0, 0x4b, 0x0a, 0xc2, 0xdf, 0xe2, 0xf0, 0x05, 0xba, 0x96,
-	0x01, 0x3e, 0x28, 0xfb, 0x67, 0x04, 0x46, 0xbc, 0x97, 0x2d, 0x7d, 0x39, 0xbe, 0x74, 0xc8, 0x15,
-	0x48, 0x73, 0xdd, 0xc2, 0x90, 0x4a, 0xe5, 0x54, 0x8b, 0x74, 0x3e, 0x42, 0xe5, 0x9d, 0x62, 0xea,
-	0xb1, 0xef, 0x88, 0x6b, 0xd1, 0xc7, 0x04, 0xae, 0xc5, 0xdb, 0x3e, 0xba, 0x9e, 0x5e, 0x33, 0xd6,
-	0xc7, 0x4a, 0x1b, 0xbd, 0x25, 0x21, 0xf6, 0x6b, 0x1c, 0x7b, 0x93, 0x6e, 0x24, 0x62, 0x77, 0x86,
-	0x00, 0x0f, 0x01, 0xdf, 0xf3, 0xff, 0x98, 0xc0, 0x73, 0x31, 0xee, 0x8c, 0xae, 0xc5, 0xb3, 0x24,
-	0x9b, 0x47, 0x29, 0xdf, 0x43, 0x06, 0xa2, 0xbf, 0xc5, 0xd1, 0xdf, 0xa4, 0xaf, 0x47, 0xd0, 0xdd,
-	0xf7, 0xbd, 0x4b, 0xdd, 0xd6, 0x9b, 0x7b, 0x92, 0xa0, 0xfe, 0xea, 0x31, 0xbf, 0xd8, 0xa2, 0x3f,
-	0x12, 0x18, 0x0f, 0x19, 0xb1, 0xa4, 0xa3, 0x36, 0xde, 0x38, 0x26, 0x1d, 0xb5, 0x09, 0xee, 0x2e,
-	0x65, 0x7e, 0xb9, 0x4f, 0xf1, 0x83, 0x87, 0x46, 0xe6, 0x1b, 0x02, 0x97, 0xfc, 0x3e, 0x28, 0xe9,
-	0xb8, 0x8d, 0x31, 0x70, 0x49, 0xc7, 0x6d, 0x9c, 0xad, 0x4a, 0x99, 0xe5, 0x36, 0x21, 0x2a, 0x8a,
-	0x1a, 0x3e, 0x22, 0x70, 0x25, 0xe8, 0x38, 0x68, 0x97, 0x13, 0x34, 0x60, 0x99, 0xa4, 0x95, 0x6c,
-	0xc1, 0x88, 0xb7, 0xce, 0xf1, 0x56, 0xe9, 0x72, 0xca, 0x79, 0x2b, 0xde, 0x08, 0xbe, 0x51, 0x3d,
-	0x21, 0x30, 0xe6, 0xf3, 0x01, 0x74, 0x21, 0xbe, 0x64, 0xd4, 0x9b, 0x48, 0x8b, 0x19, 0x22, 0x91,
-	0x6c, 0x93, 0x93, 0xad, 0x51, 0x25, 0xf9, 0x69, 0x0a, 0x4d, 0x21, 0xf7, 0x21, 0xc5, 0x37, 0x9e,
-	0x9c, 0xe6, 0xc8, 0xd3, 0xd3, 0x1c, 0xf9, 0xeb, 0x34, 0x47, 0xbe, 0x38, 0xcb, 0x0d, 0x3c, 0x3d,
-	0xcb, 0x0d, 0xfc, 0x71, 0x96, 0x1b, 0xf8, 0xe0, 0x25, 0xc3, 0x74, 0x3e, 0x3c, 0x2c, 0x2b, 0xfb,
-	0xac, 0xe6, 0xad, 0x29, 0xfe, 0x59, 0xb5, 0x2b, 0x1f, 0xa9, 0x1f, 0x8b, 0x02, 0xe5, 0x8b, 0xfc,
-	0xcf, 0xb0, 0xeb, 0xff, 0x04, 0x00, 0x00, 0xff, 0xff, 0x11, 0xfe, 0x04, 0xa9, 0x4e, 0x16, 0x00,
-	0x00,
+	// 1292 bytes of a gzipped FileDescriptorProto
+	0x1f, 0x8b, 0x08, 0x00, 0x00, 0x00, 0x00, 0x00, 0x02, 0xff, 0xcc, 0x98, 0x4b, 0x6f, 0x1b, 0x55,
+	0x14, 0xc7, 0x73, 0x4b, 0x9e, 0x27, 0x6d, 0x23, 0x6e, 0xd3, 0x36, 0x99, 0x46, 0x4e, 0x98, 0x96,
+	0xbc, 0x33, 0x13, 0x3b, 0x69, 0x5a, 0xf1, 0x54, 0x2d, 0x41, 0xc8, 0xa2, 0x28, 0x35, 0x11, 0x0b,
+	0x84, 0x14, 0x8d, 0xe3, 0x89, 0x19, 0x61, 0xcf, 0xb8, 0x9e, 0x49, 0x84, 0x15, 0x79, 0x83, 0x04,
+	0x0b, 0xc4, 0x02, 0x5a, 0x84, 0x4a, 0xc4, 0xa2, 0x0b, 0x24, 0xf8, 0x00, 0x20, 0x24, 0x76, 0xdd,
+	0x75, 0x59, 0xc1, 0x86, 0x15, 0x42, 0x09, 0xdf, 0x82, 0x0d, 0x9a, 0x7b, 0xcf, 0xb5, 0xe7, 0xed,
+	0x89, 0xb0, 0xc0, 0xab, 0x76, 0x66, 0xce, 0xb9, 0xe7, 0x37, 0xff, 0x73, 0xe6, 0xfa, 0x7f, 0x03,
+	0xd7, 0xf6, 0x2c, 0xbb, 0x6a, 0xd9, 0x6a, 0xb9, 0x6e, 0x1d, 0xd4, 0xd4, 0xc3, 0xac, 0x7a, 0xff,
+	0x40, 0xaf, 0x37, 0x94, 0x5a, 0xdd, 0x72, 0x2c, 0x3a, 0xc6, 0x1f, 0x2a, 0xec, 0xa1, 0x72, 0x98,
+	0x95, 0xc6, 0xcb, 0x56, 0xd9, 0x62, 0xcf, 0x54, 0xf7, 0x7f, 0x3c, 0x4c, 0x9a, 0x2a, 0x5b, 0x56,
+	0xb9, 0xa2, 0xab, 0x5a, 0xcd, 0x50, 0x35, 0xd3, 0xb4, 0x1c, 0xcd, 0x31, 0x2c, 0xd3, 0xc6, 0xa7,
+	0xa1, 0x0a, 0x4e, 0xa3, 0xa6, 0x8b, 0x87, 0x8b, 0xf8, 0xb0, 0xa8, 0xd9, 0x3a, 0x2f, 0xad, 0x1e,
+	0x66, 0x8b, 0xba, 0xa3, 0x65, 0xd5, 0x9a, 0x56, 0x36, 0x4c, 0xb6, 0x12, 0xc6, 0x4e, 0xf2, 0xd8,
+	0x5d, 0x5e, 0x1f, 0xd1, 0xd8, 0x85, 0x9c, 0x83, 0xcb, 0xf7, 0xdc, 0xe4, 0x4d, 0xb7, 0xc6, 0x96,
+	0xb9, 0x6f, 0x15, 0xf4, 0xfb, 0x07, 0xba, 0xed, 0xd0, 0x49, 0x18, 0x66, 0x75, 0x77, 0x8d, 0xd2,
+	0x04, 0x99, 0x21, 0xf3, 0xfd, 0x85, 0x21, 0x76, 0xbd, 0x55, 0x92, 0xdf, 0x82, 0x2b, 0xc1, 0x1c,
+	0xbb, 0x66, 0x99, 0xb6, 0x4e, 0x15, 0xe8, 0x37, 0xcc, 0x7d, 0x8b, 0x25, 0x8c, 0xe6, 0x24, 0x25,
+	0xa0, 0x82, 0xd2, 0xce, 0x60, 0x71, 0xf2, 0x3d, 0xb8, 0xd6, 0x5e, 0x69, 0xdb, 0xaa, 0x18, 0x7b,
+	0x0d, 0x2f, 0x43, 0x0e, 0x86, 0xb4, 0x52, 0xa9, 0xae, 0xdb, 0x36, 0x5b, 0x71, 0x24, 0x3f, 0xf1,
+	0xeb, 0x8f, 0x2b, 0xe3, 0xb8, 0xe8, 0x1d, 0xfe, 0xe4, 0x1d, 0xa7, 0x6e, 0x98, 0xe5, 0x82, 0x08,
+	0x94, 0x77, 0x60, 0x2a, 0x7a, 0x49, 0x44, 0x5c, 0xf7, 0x21, 0xce, 0x44, 0x23, 0x7a, 0xf2, 0x38,
+	0x68, 0x13, 0x26, 0xda, 0xab, 0xde, 0xd5, 0xab, 0x45, 0xbd, 0x6e, 0x77, 0x56, 0x8a, 0xbe, 0x09,
+	0xd0, 0x6e, 0xc6, 0xc4, 0x39, 0x56, 0x72, 0x56, 0x94, 0x74, 0x3b, 0xa7, 0xf0, 0xa1, 0xc1, 0xce,
+	0x29, 0xdb, 0x5a, 0x59, 0xc7, 0x65, 0x0b, 0x9e, 0x4c, 0xf9, 0x5b, 0x02, 0x93, 0x11, 0xf5, 0xf1,
+	0x95, 0x36, 0x60, 0xa8, 0xca, 0x6f, 0x4d, 0x90, 0x99, 0xe7, 0xe6, 0x47, 0x73, 0x53, 0xd1, 0x6f,
+	0xc5, 0xf3, 0x0a, 0x22, 0x98, 0x6e, 0x46, 0xd0, 0xcd, 0x75, 0xa4, 0xe3, 0x45, 0x7d, 0x78, 0x0f,
+	0x7d, 0x78, 0x76, 0xbe, 0x71, 0xa7, 0x54, 0x35, 0x4c, 0xa1, 0x8f, 0x02, 0x03, 0x9a, 0x7b, 0xdd,
+	0xb1, 0x87, 0x3c, 0xac, 0x6b, 0xa2, 0x7d, 0x43, 0x40, 0x8a, 0xa2, 0x42, 0xd5, 0x72, 0x30, 0xc8,
+	0xe4, 0x11, 0xa2, 0x25, 0x4d, 0x2b, 0x46, 0x76, 0x4f, 0xb1, 0x4f, 0x08, 0xcc, 0x04, 0xc6, 0xd4,
+	0xd0, 0xed, 0x3c, 0xbf, 0xfc, 0x0f, 0x07, 0xeb, 0x27, 0x02, 0x2f, 0x24, 0x70, 0xa0, 0x54, 0x9b,
+	0x70, 0x91, 0x83, 0xd4, 0x30, 0x00, 0x25, 0xeb, 0xfc, 0xf5, 0x5c, 0x28, 0x7b, 0xd7, 0xed, 0x9e,
+	0x7e, 0xc7, 0x31, 0xfa, 0xf5, 0xc4, 0xe0, 0xc5, 0x89, 0xea, 0x9f, 0xbf, 0xde, 0x13, 0xf5, 0x16,
+	0x8c, 0x33, 0xec, 0xed, 0xba, 0x55, 0xb3, 0x6c, 0xad, 0x22, 0x74, 0x9c, 0x86, 0xd1, 0x1a, 0xde,
+	0x6a, 0x8f, 0x22, 0x88, 0x5b, 0x5b, 0x25, 0xf9, 0x6d, 0xfc, 0x11, 0x69, 0x27, 0xe2, 0x3b, 0xde,
+	0x84, 0x61, 0x11, 0x86, 0x1b, 0xee, 0x64, 0xe8, 0xed, 0x5a, 0x49, 0xad, 0x50, 0xf9, 0x31, 0x01,
+	0xd9, 0xb7, 0xa0, 0x98, 0x48, 0x2e, 0xc2, 0xbf, 0xf8, 0x79, 0xe8, 0x5a, 0x8f, 0xbf, 0x27, 0x70,
+	0x3d, 0x11, 0x11, 0x15, 0xb8, 0x05, 0x23, 0xe2, 0xb5, 0x44, 0x83, 0x13, 0x24, 0x68, 0xc7, 0x76,
+	0xaf, 0xab, 0x75, 0x98, 0x66, 0xa0, 0xef, 0x5a, 0x8e, 0x9e, 0x6f, 0xe1, 0xba, 0x57, 0xf5, 0xb4,
+	0x0d, 0x76, 0xbf, 0xa4, 0x43, 0x37, 0x81, 0x71, 0x24, 0x7e, 0x49, 0x2c, 0x4c, 0xbe, 0x8b, 0x5f,
+	0x67, 0x64, 0x4d, 0x54, 0x66, 0x01, 0xfa, 0xdd, 0x60, 0x9c, 0x8b, 0xcb, 0x21, 0x51, 0xdc, 0xe8,
+	0x02, 0x0b, 0x91, 0x3f, 0x25, 0xe8, 0x13, 0xdc, 0x7b, 0x76, 0xfe, 0xcc, 0x03, 0xda, 0xb5, 0xae,
+	0x7f, 0x45, 0xd0, 0x5d, 0x84, 0x40, 0xf0, 0xa5, 0x96, 0xb8, 0x50, 0xa2, 0xd5, 0x31, 0x6f, 0xc5,
+	0x63, 0xba, 0xd7, 0xe2, 0x07, 0x04, 0xed, 0x09, 0x62, 0xf9, 0x9a, 0xdb, 0xea, 0x1d, 0x49, 0xd5,
+	0xbb, 0xae, 0x69, 0xf5, 0xa5, 0x30, 0x05, 0x7e, 0xa8, 0xff, 0x55, 0xa8, 0x47, 0x41, 0x4b, 0x80,
+	0x96, 0xa8, 0x07, 0x36, 0x94, 0x63, 0xe2, 0xf5, 0xc2, 0x1e, 0xb4, 0x5e, 0xb0, 0x2b, 0x2f, 0xc1,
+	0x55, 0xc6, 0xb6, 0xa3, 0x55, 0x2a, 0xee, 0xde, 0x76, 0x50, 0x71, 0x52, 0xff, 0x38, 0xec, 0xe0,
+	0x6c, 0xfa, 0x72, 0xf1, 0xa5, 0x6e, 0xc3, 0x80, 0xe3, 0xde, 0xc6, 0x4d, 0x20, 0xec, 0x5b, 0x3d,
+	0x49, 0xf9, 0xfe, 0xa7, 0x7f, 0x4c, 0xf7, 0x15, 0x78, 0x82, 0xfc, 0x3e, 0x50, 0x8f, 0x5a, 0x02,
+	0xa6, 0x5b, 0xcd, 0x78, 0x40, 0xe0, 0x92, 0x6f, 0xf9, 0x1e, 0x68, 0x42, 0xee, 0xef, 0xe7, 0x61,
+	0x80, 0x41, 0xd1, 0xcf, 0x09, 0x8c, 0xb4, 0x0a, 0xd1, 0xd9, 0x10, 0x44, 0xe4, 0x89, 0x4e, 0x9a,
+	0xeb, 0x18, 0xc7, 0x8b, 0xca, 0xca, 0xc7, 0xbf, 0xfd, 0xf5, 0xf0, 0xdc, 0x3c, 0x9d, 0x55, 0x83,
+	0x07, 0x50, 0xb4, 0xa3, 0xe6, 0xbe, 0xa5, 0x1e, 0x09, 0x6b, 0xda, 0xa4, 0xdf, 0x11, 0x18, 0x0b,
+	0x78, 0x14, 0xba, 0x9c, 0x50, 0x2c, 0x74, 0xd0, 0x93, 0x56, 0x52, 0x46, 0x23, 0xe0, 0x3a, 0x03,
+	0x54, 0xe8, 0x72, 0x0c, 0x20, 0x73, 0x54, 0x0d, 0xe4, 0xc4, 0x0f, 0xb5, 0x49, 0x1f, 0x11, 0x38,
+	0xef, 0x3d, 0x3f, 0xd1, 0x85, 0x84, 0xaa, 0xfe, 0x33, 0x9e, 0xb4, 0x98, 0x26, 0x14, 0xe9, 0xb2,
+	0x8c, 0x6e, 0x89, 0x2e, 0xc4, 0xd0, 0xe1, 0xf1, 0xcb, 0xab, 0xe0, 0x31, 0x81, 0x0b, 0xbe, 0x53,
+	0x0a, 0x4d, 0x2a, 0x18, 0xf0, 0xb9, 0xd2, 0x52, 0xaa, 0x58, 0xa4, 0x5b, 0x65, 0x74, 0x8b, 0x74,
+	0x3e, 0x9a, 0xce, 0xde, 0x2d, 0x36, 0x76, 0x99, 0x1d, 0x76, 0x95, 0xab, 0x1a, 0x66, 0x93, 0xfe,
+	0x42, 0x60, 0x3c, 0xea, 0x78, 0x40, 0xb3, 0x9d, 0xba, 0x16, 0x3a, 0xd2, 0x48, 0xb9, 0xb3, 0xa4,
+	0x20, 0xf1, 0xcb, 0x8c, 0xf8, 0x26, 0x5d, 0x4b, 0xea, 0xb6, 0xa1, 0x33, 0x72, 0xfe, 0xc8, 0xa3,
+	0xec, 0xcf, 0x61, 0x78, 0x2e, 0x70, 0x3a, 0x78, 0x9f, 0xce, 0xb9, 0xb3, 0xa4, 0x20, 0xfc, 0x6d,
+	0x06, 0x9f, 0xa3, 0xab, 0x29, 0xe0, 0xfd, 0xb2, 0x7f, 0x46, 0x60, 0x58, 0xf8, 0x0b, 0xfa, 0x62,
+	0x74, 0xe9, 0x80, 0x11, 0x92, 0x66, 0x3b, 0x85, 0x21, 0x95, 0xca, 0xa8, 0x16, 0xe8, 0x5c, 0x88,
+	0x4a, 0x6c, 0xdc, 0xea, 0x91, 0x67, 0x57, 0x6f, 0xd2, 0x27, 0x04, 0xae, 0x44, 0x3b, 0x5d, 0xba,
+	0x96, 0x5c, 0x33, 0xd2, 0xba, 0x4b, 0xeb, 0x67, 0x4b, 0x42, 0xec, 0x57, 0x18, 0xf6, 0x06, 0x5d,
+	0x8f, 0xc5, 0x6e, 0x0f, 0x01, 0x6e, 0x02, 0x9e, 0xef, 0xff, 0x09, 0x81, 0x4b, 0x11, 0x86, 0x94,
+	0xae, 0x46, 0xb3, 0xc4, 0xfb, 0x65, 0x29, 0x7b, 0x86, 0x0c, 0x44, 0x7f, 0x83, 0xa1, 0xbf, 0x4e,
+	0x5f, 0x0d, 0xa1, 0xbb, 0x16, 0xc7, 0xa5, 0x6e, 0xe9, 0xcd, 0x6c, 0x98, 0x5f, 0x7f, 0xf5, 0x88,
+	0xdd, 0x6c, 0xd2, 0x1f, 0x08, 0x8c, 0x05, 0xbc, 0x67, 0xdc, 0x56, 0x1b, 0xed, 0x95, 0xe3, 0xb6,
+	0xda, 0x18, 0x43, 0x9b, 0x30, 0xbf, 0xcc, 0x9a, 0x79, 0xc1, 0x03, 0x23, 0xf3, 0x35, 0x81, 0xf3,
+	0x5e, 0xeb, 0x17, 0xb7, 0xdd, 0x46, 0x78, 0xd6, 0xb8, 0xed, 0x36, 0xca, 0x49, 0x26, 0xcc, 0x72,
+	0x8b, 0x10, 0x15, 0x45, 0x0d, 0x1f, 0x13, 0xb8, 0xe8, 0x37, 0x59, 0xb4, 0xc3, 0x0e, 0xea, 0x73,
+	0x89, 0xd2, 0x72, 0xba, 0x60, 0xc4, 0x5b, 0x63, 0x78, 0x2b, 0x74, 0x29, 0x61, 0xbf, 0xe5, 0xbf,
+	0x08, 0x9e, 0x51, 0x3d, 0x26, 0x30, 0xea, 0xb1, 0x3e, 0x74, 0x3e, 0xba, 0x64, 0xd8, 0x8e, 0x49,
+	0x0b, 0x29, 0x22, 0x91, 0x6c, 0x83, 0x91, 0xad, 0x52, 0x25, 0xfe, 0x6b, 0x0a, 0x4c, 0x21, 0xb3,
+	0x5e, 0xd4, 0x81, 0x41, 0xfe, 0xae, 0xf4, 0x7a, 0x92, 0x12, 0x82, 0xe8, 0x46, 0x72, 0x10, 0xc2,
+	0x4c, 0x33, 0x98, 0x49, 0x7a, 0x35, 0x46, 0xa6, 0xfc, 0x6b, 0x4f, 0x4f, 0x32, 0xe4, 0xd9, 0x49,
+	0x86, 0xfc, 0x79, 0x92, 0x21, 0x5f, 0x9c, 0x66, 0xfa, 0x9e, 0x9d, 0x66, 0xfa, 0x7e, 0x3f, 0xcd,
+	0xf4, 0xbd, 0x77, 0xa3, 0x6c, 0x38, 0x1f, 0x1c, 0x14, 0x95, 0x3d, 0xab, 0x2a, 0x92, 0xf9, 0x3f,
+	0x2b, 0x76, 0xe9, 0x43, 0xf5, 0x23, 0xbe, 0x40, 0x71, 0x90, 0xfd, 0xbd, 0x7b, 0xed, 0x9f, 0x00,
+	0x00, 0x00, 0xff, 0xff, 0x1c, 0xce, 0x79, 0xa7, 0xb7, 0x17, 0x00, 0x00,
 }
 
 // Reference imports to suppress errors if they are not otherwise used.
@@ -1536,6 +1645,10 @@ type QueryClient interface {
 	// then it simply returns the `final_tally_result` state stored in the
 	// proposal itself.
 	TallyResult(ctx context.Context, in *QueryTallyResultRequest, opts ...grpc.CallOption) (*QueryTallyResultResponse, error)
+	// Groups queries all groups in state.
+	//
+	// Since: cosmos-sdk 0.47.1
+	Groups(ctx context.Context, in *QueryGroupsRequest, opts ...grpc.CallOption) (*QueryGroupsResponse, error)
 }
 
 type queryClient struct {
@@ -1663,6 +1776,15 @@ func (c *queryClient) TallyResult(ctx context.Context, in *QueryTallyResultReque
 	return out, nil
 }
 
+func (c *queryClient) Groups(ctx context.Context, in *QueryGroupsRequest, opts ...grpc.CallOption) (*QueryGroupsResponse, error) {
+	out := new(QueryGroupsResponse)
+	err := c.cc.Invoke(ctx, "/cosmos.group.v1.Query/Groups", in, out, opts...)
+	if err != nil {
+		return nil, err
+	}
+	return out, nil
+}
+
 // QueryServer is the server API for Query service.
 type QueryServer interface {
 	// GroupInfo queries group info based on group id.
@@ -1695,6 +1817,10 @@ type QueryServer interface {
 	// then it simply returns the `final_tally_result` state stored in the
 	// proposal itself.
 	TallyResult(context.Context, *QueryTallyResultRequest) (*QueryTallyResultResponse, error)
+	// Groups queries all groups in state.
+	//
+	// Since: cosmos-sdk 0.47.1
+	Groups(context.Context, *QueryGroupsRequest) (*QueryGroupsResponse, error)
 }
 
 // UnimplementedQueryServer can be embedded to have forward compatible implementations.
@@ -1739,6 +1865,9 @@ func (*UnimplementedQueryServer) GroupsByMember(ctx context.Context, req *QueryG
 }
 func (*UnimplementedQueryServer) TallyResult(ctx context.Context, req *QueryTallyResultRequest) (*QueryTallyResultResponse, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method TallyResult not implemented")
+}
+func (*UnimplementedQueryServer) Groups(ctx context.Context, req *QueryGroupsRequest) (*QueryGroupsResponse, error) {
+	return nil, status.Errorf(codes.Unimplemented, "method Groups not implemented")
 }
 
 func RegisterQueryServer(s grpc1.Server, srv QueryServer) {
@@ -1979,6 +2108,24 @@ func _Query_TallyResult_Handler(srv interface{}, ctx context.Context, dec func(i
 	return interceptor(ctx, in, info, handler)
 }
 
+func _Query_Groups_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
+	in := new(QueryGroupsRequest)
+	if err := dec(in); err != nil {
+		return nil, err
+	}
+	if interceptor == nil {
+		return srv.(QueryServer).Groups(ctx, in)
+	}
+	info := &grpc.UnaryServerInfo{
+		Server:     srv,
+		FullMethod: "/cosmos.group.v1.Query/Groups",
+	}
+	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
+		return srv.(QueryServer).Groups(ctx, req.(*QueryGroupsRequest))
+	}
+	return interceptor(ctx, in, info, handler)
+}
+
 var _Query_serviceDesc = grpc.ServiceDesc{
 	ServiceName: "cosmos.group.v1.Query",
 	HandlerType: (*QueryServer)(nil),
@@ -2034,6 +2181,10 @@ var _Query_serviceDesc = grpc.ServiceDesc{
 		{
 			MethodName: "TallyResult",
 			Handler:    _Query_TallyResult_Handler,
+		},
+		{
+			MethodName: "Groups",
+			Handler:    _Query_Groups_Handler,
 		},
 	},
 	Streams:  []grpc.StreamDesc{},
@@ -3084,6 +3235,90 @@ func (m *QueryTallyResultResponse) MarshalToSizedBuffer(dAtA []byte) (int, error
 	return len(dAtA) - i, nil
 }
 
+func (m *QueryGroupsRequest) Marshal() (dAtA []byte, err error) {
+	size := m.Size()
+	dAtA = make([]byte, size)
+	n, err := m.MarshalToSizedBuffer(dAtA[:size])
+	if err != nil {
+		return nil, err
+	}
+	return dAtA[:n], nil
+}
+
+func (m *QueryGroupsRequest) MarshalTo(dAtA []byte) (int, error) {
+	size := m.Size()
+	return m.MarshalToSizedBuffer(dAtA[:size])
+}
+
+func (m *QueryGroupsRequest) MarshalToSizedBuffer(dAtA []byte) (int, error) {
+	i := len(dAtA)
+	_ = i
+	var l int
+	_ = l
+	if m.Pagination != nil {
+		{
+			size, err := m.Pagination.MarshalToSizedBuffer(dAtA[:i])
+			if err != nil {
+				return 0, err
+			}
+			i -= size
+			i = encodeVarintQuery(dAtA, i, uint64(size))
+		}
+		i--
+		dAtA[i] = 0x12
+	}
+	return len(dAtA) - i, nil
+}
+
+func (m *QueryGroupsResponse) Marshal() (dAtA []byte, err error) {
+	size := m.Size()
+	dAtA = make([]byte, size)
+	n, err := m.MarshalToSizedBuffer(dAtA[:size])
+	if err != nil {
+		return nil, err
+	}
+	return dAtA[:n], nil
+}
+
+func (m *QueryGroupsResponse) MarshalTo(dAtA []byte) (int, error) {
+	size := m.Size()
+	return m.MarshalToSizedBuffer(dAtA[:size])
+}
+
+func (m *QueryGroupsResponse) MarshalToSizedBuffer(dAtA []byte) (int, error) {
+	i := len(dAtA)
+	_ = i
+	var l int
+	_ = l
+	if m.Pagination != nil {
+		{
+			size, err := m.Pagination.MarshalToSizedBuffer(dAtA[:i])
+			if err != nil {
+				return 0, err
+			}
+			i -= size
+			i = encodeVarintQuery(dAtA, i, uint64(size))
+		}
+		i--
+		dAtA[i] = 0x12
+	}
+	if len(m.Groups) > 0 {
+		for iNdEx := len(m.Groups) - 1; iNdEx >= 0; iNdEx-- {
+			{
+				size, err := m.Groups[iNdEx].MarshalToSizedBuffer(dAtA[:i])
+				if err != nil {
+					return 0, err
+				}
+				i -= size
+				i = encodeVarintQuery(dAtA, i, uint64(size))
+			}
+			i--
+			dAtA[i] = 0xa
+		}
+	}
+	return len(dAtA) - i, nil
+}
+
 func encodeVarintQuery(dAtA []byte, offset int, v uint64) int {
 	offset -= sovQuery(v)
 	base := offset
@@ -3505,6 +3740,38 @@ func (m *QueryTallyResultResponse) Size() (n int) {
 	_ = l
 	l = m.Tally.Size()
 	n += 1 + l + sovQuery(uint64(l))
+	return n
+}
+
+func (m *QueryGroupsRequest) Size() (n int) {
+	if m == nil {
+		return 0
+	}
+	var l int
+	_ = l
+	if m.Pagination != nil {
+		l = m.Pagination.Size()
+		n += 1 + l + sovQuery(uint64(l))
+	}
+	return n
+}
+
+func (m *QueryGroupsResponse) Size() (n int) {
+	if m == nil {
+		return 0
+	}
+	var l int
+	_ = l
+	if len(m.Groups) > 0 {
+		for _, e := range m.Groups {
+			l = e.Size()
+			n += 1 + l + sovQuery(uint64(l))
+		}
+	}
+	if m.Pagination != nil {
+		l = m.Pagination.Size()
+		n += 1 + l + sovQuery(uint64(l))
+	}
 	return n
 }
 
@@ -6172,6 +6439,212 @@ func (m *QueryTallyResultResponse) Unmarshal(dAtA []byte) error {
 				return io.ErrUnexpectedEOF
 			}
 			if err := m.Tally.Unmarshal(dAtA[iNdEx:postIndex]); err != nil {
+				return err
+			}
+			iNdEx = postIndex
+		default:
+			iNdEx = preIndex
+			skippy, err := skipQuery(dAtA[iNdEx:])
+			if err != nil {
+				return err
+			}
+			if (skippy < 0) || (iNdEx+skippy) < 0 {
+				return ErrInvalidLengthQuery
+			}
+			if (iNdEx + skippy) > l {
+				return io.ErrUnexpectedEOF
+			}
+			iNdEx += skippy
+		}
+	}
+
+	if iNdEx > l {
+		return io.ErrUnexpectedEOF
+	}
+	return nil
+}
+func (m *QueryGroupsRequest) Unmarshal(dAtA []byte) error {
+	l := len(dAtA)
+	iNdEx := 0
+	for iNdEx < l {
+		preIndex := iNdEx
+		var wire uint64
+		for shift := uint(0); ; shift += 7 {
+			if shift >= 64 {
+				return ErrIntOverflowQuery
+			}
+			if iNdEx >= l {
+				return io.ErrUnexpectedEOF
+			}
+			b := dAtA[iNdEx]
+			iNdEx++
+			wire |= uint64(b&0x7F) << shift
+			if b < 0x80 {
+				break
+			}
+		}
+		fieldNum := int32(wire >> 3)
+		wireType := int(wire & 0x7)
+		if wireType == 4 {
+			return fmt.Errorf("proto: QueryGroupsRequest: wiretype end group for non-group")
+		}
+		if fieldNum <= 0 {
+			return fmt.Errorf("proto: QueryGroupsRequest: illegal tag %d (wire type %d)", fieldNum, wire)
+		}
+		switch fieldNum {
+		case 2:
+			if wireType != 2 {
+				return fmt.Errorf("proto: wrong wireType = %d for field Pagination", wireType)
+			}
+			var msglen int
+			for shift := uint(0); ; shift += 7 {
+				if shift >= 64 {
+					return ErrIntOverflowQuery
+				}
+				if iNdEx >= l {
+					return io.ErrUnexpectedEOF
+				}
+				b := dAtA[iNdEx]
+				iNdEx++
+				msglen |= int(b&0x7F) << shift
+				if b < 0x80 {
+					break
+				}
+			}
+			if msglen < 0 {
+				return ErrInvalidLengthQuery
+			}
+			postIndex := iNdEx + msglen
+			if postIndex < 0 {
+				return ErrInvalidLengthQuery
+			}
+			if postIndex > l {
+				return io.ErrUnexpectedEOF
+			}
+			if m.Pagination == nil {
+				m.Pagination = &query.PageRequest{}
+			}
+			if err := m.Pagination.Unmarshal(dAtA[iNdEx:postIndex]); err != nil {
+				return err
+			}
+			iNdEx = postIndex
+		default:
+			iNdEx = preIndex
+			skippy, err := skipQuery(dAtA[iNdEx:])
+			if err != nil {
+				return err
+			}
+			if (skippy < 0) || (iNdEx+skippy) < 0 {
+				return ErrInvalidLengthQuery
+			}
+			if (iNdEx + skippy) > l {
+				return io.ErrUnexpectedEOF
+			}
+			iNdEx += skippy
+		}
+	}
+
+	if iNdEx > l {
+		return io.ErrUnexpectedEOF
+	}
+	return nil
+}
+func (m *QueryGroupsResponse) Unmarshal(dAtA []byte) error {
+	l := len(dAtA)
+	iNdEx := 0
+	for iNdEx < l {
+		preIndex := iNdEx
+		var wire uint64
+		for shift := uint(0); ; shift += 7 {
+			if shift >= 64 {
+				return ErrIntOverflowQuery
+			}
+			if iNdEx >= l {
+				return io.ErrUnexpectedEOF
+			}
+			b := dAtA[iNdEx]
+			iNdEx++
+			wire |= uint64(b&0x7F) << shift
+			if b < 0x80 {
+				break
+			}
+		}
+		fieldNum := int32(wire >> 3)
+		wireType := int(wire & 0x7)
+		if wireType == 4 {
+			return fmt.Errorf("proto: QueryGroupsResponse: wiretype end group for non-group")
+		}
+		if fieldNum <= 0 {
+			return fmt.Errorf("proto: QueryGroupsResponse: illegal tag %d (wire type %d)", fieldNum, wire)
+		}
+		switch fieldNum {
+		case 1:
+			if wireType != 2 {
+				return fmt.Errorf("proto: wrong wireType = %d for field Groups", wireType)
+			}
+			var msglen int
+			for shift := uint(0); ; shift += 7 {
+				if shift >= 64 {
+					return ErrIntOverflowQuery
+				}
+				if iNdEx >= l {
+					return io.ErrUnexpectedEOF
+				}
+				b := dAtA[iNdEx]
+				iNdEx++
+				msglen |= int(b&0x7F) << shift
+				if b < 0x80 {
+					break
+				}
+			}
+			if msglen < 0 {
+				return ErrInvalidLengthQuery
+			}
+			postIndex := iNdEx + msglen
+			if postIndex < 0 {
+				return ErrInvalidLengthQuery
+			}
+			if postIndex > l {
+				return io.ErrUnexpectedEOF
+			}
+			m.Groups = append(m.Groups, &GroupInfo{})
+			if err := m.Groups[len(m.Groups)-1].Unmarshal(dAtA[iNdEx:postIndex]); err != nil {
+				return err
+			}
+			iNdEx = postIndex
+		case 2:
+			if wireType != 2 {
+				return fmt.Errorf("proto: wrong wireType = %d for field Pagination", wireType)
+			}
+			var msglen int
+			for shift := uint(0); ; shift += 7 {
+				if shift >= 64 {
+					return ErrIntOverflowQuery
+				}
+				if iNdEx >= l {
+					return io.ErrUnexpectedEOF
+				}
+				b := dAtA[iNdEx]
+				iNdEx++
+				msglen |= int(b&0x7F) << shift
+				if b < 0x80 {
+					break
+				}
+			}
+			if msglen < 0 {
+				return ErrInvalidLengthQuery
+			}
+			postIndex := iNdEx + msglen
+			if postIndex < 0 {
+				return ErrInvalidLengthQuery
+			}
+			if postIndex > l {
+				return io.ErrUnexpectedEOF
+			}
+			if m.Pagination == nil {
+				m.Pagination = &query.PageResponse{}
+			}
+			if err := m.Pagination.Unmarshal(dAtA[iNdEx:postIndex]); err != nil {
 				return err
 			}
 			iNdEx = postIndex
